// src/lib/api/models/wasteReportListItem.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * FastAPI JWT Auth Boilerplate
 * FastAPI Boilerplate with JWT Authentication
 * OpenAPI spec version: 0.1.0
 */
import type { WasteType } from './wasteType';
import type { UrgencyLevel } from './urgencyLevel';
import type { ReportStatus } from './reportStatus';

export interface WasteReportListItem {
  id: number;
  street_address: string;
  city: string;
  waste_type: WasteType;
  urgency_level: UrgencyLevel;
  status: ReportStatus;
  created_at: string;
  photo_count?: number;
}