// src/lib/api/models/urgencyLevel.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * FastAPI JWT Auth Boilerplate
 * FastAPI Boilerplate with JWT Authentication
 * OpenAPI spec version: 0.1.0
 */

export type UrgencyLevel = typeof UrgencyLevel[keyof typeof UrgencyLevel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UrgencyLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;