// src/lib/api/models/reportStatus.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * FastAPI JWT Auth Boilerplate
 * FastAPI Boilerplate with JWT Authentication
 * OpenAPI spec version: 0.1.0
 */

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  completed: 'completed',
} as const;